backend/requirements.txt
Django==5.2.3
asgiref==3.8.1
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
django-cors-headers==4.7.0
mysqlclient==2.2.7
Pillow==11.2.1
qrcode==8.2
django-reversion==5.1.0
decouple==0.0.7
python-decouple==3.8
